Dockerized Flask app/requirements.txt
db-dtypes==1.4.2
Flask==2.3.2
google-cloud-bigquery==3.12.0
pandas
